Refactor more parts of the transaction sending process
