+++ client/app/src/components/account/templates/send-transaction-dialog.html
@@ -1,8 +1,8 @@
-<md-dialog aria-label="{{'Send'|translate}} {{ul.network.token}}" ng-cloak md-theme="ul.currentTheme">
-  <form name="sendArkForm" ng-submit="send.next()">
+<md-dialog aria-label="{{'Send'|translate}} {{ac.network.token}}" ng-cloak md-theme="ac.currentTheme">
+  <form name="sendArkForm" ng-submit="send.submit()">
     <md-toolbar>
       <div class="md-toolbar-tools">
-        <h2><span>{{'Send'|translate}} {{ul.network.token}} {{'from'|translate}}</span>&nbsp;{{send.data.fromAddress}}</h2>
+        <h2><span>{{'Send'|translate}} {{ac.network.token}} {{'from'|translate}}</span>&nbsp;{{send.data.fromAddress}}</h2>
         <span flex></span>
         <md-button class="md-icon-button" ng-click="send.cancel()">
           <md-icon md-font-library="material-icons">clear</md-icon>
@@ -26,10 +26,10 @@
         </div>
         <div layout-gt-sm="row">
           <md-input-container md-no-float class="md-block" flex-gt-sm="85">
-            <label>{{'Amount'|translate}} ({{ul.network.token}})</label>
+            <label>{{'Amount'|translate}} ({{ac.network.token}})</label>
             <input ng-model="send.data.amount" type="number" step="0.00000001" min="0.00000001" max="100000000000000" valid-amount autocomplete="off" ng-required="true" tabIndex="2">
             <div class="hint">
-              {{ul.currency.symbol}} {{send.data.amount | amountToCurrency:this}}
+              {{ac.currency.symbol}} {{send.data.amount | amountToCurrency:this}}
             </div>
           </md-input-container>
           <md-input-container class="md-block" flex-gt-sm="15">
@@ -74,7 +74,7 @@
         <div class="md-caption">
           <translate>Remaining balance</translate>
         </div>
-        <div class="balance">{{ul.network.symbol}} {{ send.remainingBalance }}</div>
+        <div class="balance">{{ac.network.symbol}} {{ send.remainingBalance }}</div>
       </div>
       <span flex></span>
       <md-button ng-click="send.openFile()">
